refactor(plugin) move files
