+++ lua/plugins/init.lua
@@ -38,15 +38,17 @@ require('packer').startup(function(use)
 end)
 
 -- plugin setups
-require('plugins.autopairs')
-require('plugins.bufferline')
-require('plugins.comment')
-require('plugins.indentline')
-require('plugins.lualine')
-require('plugins.luasnip')
-require('plugins.surround')
-require('plugins.telescope')
-require('plugins.treesitter')
-require('plugins.whichkey')
-require('plugins.external.mason')
-require('plugins.cmp')
+require('plugins/external/mason')
+
+require('plugins/behavior/autopairs')
+require('plugins/behavior/comment')
+require('plugins/behavior/luasnip')
+require('plugins/behavior/surround')
+require('plugins/behavior/telescope')
+require('plugins/behavior/cmp')
+
+require('plugins/interface/bufferline')
+require('plugins/interface/indentline')
+require('plugins/interface/lualine')
+require('plugins/interface/treesitter')
+require('plugins/interface/whichkey')
